Better modified lines highlighting
